paranuara_project/requirements.txt
Django==1.11.4
djangorestframework==3.6.3
django-taggit==0.22.1
python-dateutil==2.6.1
